+++ src/app/ui/SignupSection.jsx
@@ -1,21 +1,27 @@
 "use client";
 import { Button } from "primereact/button";
 
-export default function SignupSection({ signUpPossible }) {
-  const content =
-    "Did we peak your enthusiasm? Great! You can sign up with your team. Don't have a team yet? Don't worry, we can match you to an awesome team! If the current edition has already started, you can sign up for the next one.";
+export default function SignupSection({ data }) {
 
-  function signUpHandler() {
+  function handleSignUp() {
     console.log("Clicked!");
   }
 
   return (
     <section className="text-center">
-      <p className="w-2/3 m-auto mb-4">{content}</p>
-      {signUpPossible ? (
-        <Button label="Sign up" onClick={signUpHandler}></Button>
+      <p className="w-2/3 m-auto mb-4">{data.content}</p>
+      {data.signUpPossible ? (
+        <Button 
+          label={data.buttonLabels.label1} 
+          onClick={handleSignUp} 
+          disabled={!data.signUpPossible}
+        ></Button>
       ) : (
-        <Button label="Sign up next year!" onClick={signUpHandler}></Button>
+        <Button 
+          label={data.buttonLabels.label2} 
+          onClick={handleSignUp} 
+          disabled={!data.signUpPossible}
+        ></Button>
       )}
     </section>
   );